+++ example/.umirc.ts
@@ -5,13 +5,23 @@ export default defineConfig({
     require.resolve('../packages/preset-react/lib'),
     // require.resolve('../packages/plugin-esbuild'),
   ],
-  plugins: [require.resolve('../packages/plugin-esbuild/lib')],
+  plugins: [
+    require.resolve('../packages/plugin-esbuild/lib'),
+    require.resolve('../packages/plugin-qiankun/lib'),
+    require.resolve('../packages/plugin-antd-dayjs/lib'),
+  ],
   routes: [
     {
       name: 'model 测试',
       path: '/plugin-model',
       component: './plugin-model',
       icon: 'smile',
+    },
+    {
+      name: 'qiankun 测试',
+      path: '/plugin-qiankun',
+      component: './plugin-qiankun',
+      icon: 'setting',
     },
     {
       name: 'initial-state 测试',
@@ -53,5 +63,18 @@ export default defineConfig({
   locale: {},
   layout: {
     name: 'UMI 3',
+  },
+  antd: {
+    config: {},
+  },
+  qiankun: {
+    master: {
+      apps: [
+        {
+          name: 'taobao',
+          entry: 'https://taobao.com',
+        },
+      ],
+    },
   },
 });