+++ packages/plugin-access/src/fixtures/normal/.umirc.ts
@@ -7,5 +7,6 @@ export default {
     require.resolve('../../'),
     require.resolve('@umijs/plugin-initial-state'),
     require.resolve('@umijs/plugin-model'),
-  ]
+  ],
+  access: {}
 }