+++ packages/plugin-access/src/fixtures/normal/pages/index.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 const App = (props: any) => {
   return (
     <div>
-      <h1>unaccessible: {props.route.unaccessible.toString()}</h1>
+      <h1>unaccessible: {props.route.unaccessible?.toString()}</h1>
     </div>
   );
 };