+++ packages/plugin-request/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.8.0](https://github.com/umijs/plugins/compare/@umijs/plugin-request@2.7.0...@umijs/plugin-request@2.8.0) (2021-09-13)
+
+### Features
+
+- BizError 返回 response ([#711](https://github.com/umijs/plugins/issues/711)) ([f15b4c8](https://github.com/umijs/plugins/commit/f15b4c89a3c709c375bcfa906a21ffcd4e715111))
+
+# [2.7.0](https://github.com/umijs/plugins/compare/@umijs/plugin-request@2.6.0...@umijs/plugin-request@2.7.0) (2021-07-21)
+
+### Features
+
+- **request:** notification support description ([3115bd0](https://github.com/umijs/plugins/commit/3115bd0f14d57e3091e3ffe9f84332d65d655e67))
+
 # [2.6.0](https://github.com/umijs/plugins/compare/@umijs/plugin-request@2.5.2...@umijs/plugin-request@2.6.0) (2021-06-03)
 
 ### Features