+++ packages/preset-react/CHANGELOG.md
@@ -2,6 +2,78 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.30](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.29...@umijs/preset-react@1.8.30) (2021-11-23)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.29](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.28...@umijs/preset-react@1.8.29) (2021-11-17)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.28](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.27...@umijs/preset-react@1.8.28) (2021-11-16)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.27](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.26...@umijs/preset-react@1.8.27) (2021-11-13)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.26](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.25...@umijs/preset-react@1.8.26) (2021-11-12)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.25](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.24...@umijs/preset-react@1.8.25) (2021-10-26)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.24](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.23...@umijs/preset-react@1.8.24) (2021-09-13)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.23](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.22...@umijs/preset-react@1.8.23) (2021-09-06)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.22](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.21...@umijs/preset-react@1.8.22) (2021-07-29)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.21](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.20...@umijs/preset-react@1.8.21) (2021-07-21)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.20](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.19...@umijs/preset-react@1.8.20) (2021-07-20)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.19](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.18...@umijs/preset-react@1.8.19) (2021-07-16)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.18](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.17...@umijs/preset-react@1.8.18) (2021-07-16)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.17](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.16...@umijs/preset-react@1.8.17) (2021-07-13)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.16](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.15...@umijs/preset-react@1.8.16) (2021-07-13)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.15](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.14...@umijs/preset-react@1.8.15) (2021-07-13)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.14](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.13...@umijs/preset-react@1.8.14) (2021-06-30)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.8.13](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.12...@umijs/preset-react@1.8.13) (2021-06-23)
+
+**Note:** Version bump only for package @umijs/preset-react
+
 ## [1.8.12](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.8.11...@umijs/preset-react@1.8.12) (2021-06-07)
 
 **Note:** Version bump only for package @umijs/preset-react